CdmsGateway.Test/Services/Routing/MessageRoutesTests.cs
using CdmsGateway.Services.Routing;
using FluentAssertions;
using NSubstitute;
using Serilog;

namespace CdmsGateway.Test.Services.Routing;

public class MessageRoutesTests
{
    [Theory]
    [InlineData("/route-1/sub-path/", "http://url-1/sub-path")]
    [InlineData("/route-2/sub-path/", "http://url-2/sub-path")]
    [InlineData("/route-3/sub-path/", "http://url-3/sub-path")]
    [InlineData("/route-4/sub-path/", "http://url-1/sub-path")]
    public void When_routing_returned_route_Then_should_get_correct_route(string routedPath, string expectedRoutePath)
    {
        var messageRoutes = new MessageRoutes(TestRoutes.RoutingConfig, Substitute.For<ILogger>());

        var route = messageRoutes.GetReturnedRoute(routedPath);

        route.RouteFound.Should().BeTrue();
        route.RouteUrl.Should().Be(expectedRoutePath);
    }

    [Theory]
    [InlineData("/route-1/sub-path/", "http://url-2/sub-path")]
    [InlineData("/route-2/sub-path/", "http://url-3/sub-path")]
    [InlineData("/route-3/sub-path/", "http://url-4/sub-path")]
    [InlineData("/route-4/sub-path/", "http://url-3/sub-path")]
    public void When_routing_unreturned_route_Then_should_get_correct_route(string routedPath, string expectedRoutePath)
    {
        var messageRoutes = new MessageRoutes(TestRoutes.RoutingConfig, Substitute.For<ILogger>());

        var route = messageRoutes.GetUnreturnedRoute(routedPath);

        route.RouteFound.Should().BeTrue();
        route.RouteUrl.Should().Be(expectedRoutePath);
    }
}